combine file view and vre view
