+++ src/components/SearchResultItem.tsx
@@ -52,21 +52,6 @@ export const SearchResultItem = ({hit, isAiRanked = false}: SearchResultItemProp
         window.open(`/dataplayer_legacy_run?${params.toString()}`, '_blank');
     };
 
-    const handleDataplayerNewRun = () => {
-        // Open the dispatcher run page in a new tab with dataset info
-        const params = new URLSearchParams();
-        params.set('datasetId', hit._id);
-        if (hit.title) {
-            params.set('title', hit.title);
-        }
-        // Preserve the search query for back navigation
-        const currentQuery = searchParams.get('q');
-        if (currentQuery) {
-            params.set('q', currentQuery);
-        }
-        window.open(`/dataplayer_new_run?${params.toString()}`, '_blank');
-    };
-
     const scorePercent = (hit.score || 0) * 100;
 
     const getPublicationDate = (): string | null => {
@@ -121,7 +106,7 @@ export const SearchResultItem = ({hit, isAiRanked = false}: SearchResultItemProp
                 </h3>
                 {isAiRanked && hit.score !== undefined && (
                     <div
-                        className="flex-shrink-0 flex items-center space-x-1 bg-yellow-50 px-2 py-1 rounded-full cursor-help"
+                        className="shrink-0 flex items-center space-x-1 bg-yellow-50 px-2 py-1 rounded-full cursor-help"
                         title="AI-powered relevance score: Ranked by LLM based on semantic understanding of your query. Scores range from 0% (low relevance) to 100% (high relevance)."
                     >
                         <ProportionalStar percent={scorePercent} className="h-4 w-4"/>
@@ -132,7 +117,7 @@ export const SearchResultItem = ({hit, isAiRanked = false}: SearchResultItemProp
                 )}
                 {!isAiRanked && typeof hit._score === 'number' && !Number.isNaN(hit._score) && (
                     <div
-                        className="flex-shrink-0 flex items-center space-x-1 bg-blue-50 px-2 py-1 rounded-full cursor-help"
+                        className="shrink-0 flex items-center space-x-1 bg-blue-50 px-2 py-1 rounded-full cursor-help"
                         title="OpenSearch relevance score: Based on keyword matching and text analysis. Scores range from 0% (low match) to 100% (high match)."
                     >
                         <ProportionalStar percent={(hit._score || 0) * 100} className="h-4 w-4" color="#005EB8"/>
@@ -162,7 +147,7 @@ export const SearchResultItem = ({hit, isAiRanked = false}: SearchResultItemProp
             <div className="space-y-2 mb-4">
                 {creators.length > 0 && (
                     <div className="flex items-start space-x-2">
-                        <UserIcon className="h-4 w-4 text-gray-500 flex-shrink-0 mt-0.5"/>
+                        <UserIcon className="h-4 w-4 text-gray-500 shrink-0 mt-0.5"/>
                         <div>
                             <span className="text-sm text-gray-600">
                                 {visibleCreators.map(c => c.creatorName).join(', ')}
@@ -185,7 +170,7 @@ export const SearchResultItem = ({hit, isAiRanked = false}: SearchResultItemProp
 
                 {publicationDate && (
                     <div className="flex items-center space-x-2">
-                        <CalendarIcon className="h-4 w-4 text-gray-500 flex-shrink-0"/>
+                        <CalendarIcon className="h-4 w-4 text-gray-500 shrink-0"/>
                         <span className="text-sm text-gray-600">
                             {formatDate(publicationDate)}
                         </span>
@@ -194,11 +179,11 @@ export const SearchResultItem = ({hit, isAiRanked = false}: SearchResultItemProp
 
                 {hit._source.subjects && hit._source.subjects.length > 0 && (
                     <div className="flex items-start space-x-2">
-                        <TagIcon className="h-4 w-4 text-gray-500 flex-shrink-0 mt-0.5"/>
+                        <TagIcon className="h-4 w-4 text-gray-500 shrink-0 mt-0.5"/>
                         <div className="flex flex-wrap gap-1">
                             {hit._source.subjects.slice(0, 5).map((subj, index) => (
                                 <span key={index}
-                                      className="inline-block bg-blue-50 text-blue-700 text-xs px-2 py-1 rounded-full">
+                                    className="inline-block bg-blue-50 text-blue-700 text-xs px-2 py-1 rounded-full">
                                     {subj.subject}
                                 </span>
                             ))}
@@ -223,13 +208,6 @@ export const SearchResultItem = ({hit, isAiRanked = false}: SearchResultItemProp
                         <span className="leading-none">Run (to be deprecated)</span>
                     </button>
                     <button
-                        onClick={handleDataplayerNewRun}
-                        aria-label={`Play dataset for ${hit.title}`}
-                        className="inline-flex items-center justify-center gap-1 rounded-md bg-blue-600 px-3 py-1.5 text-sm font-medium text-white shadow-sm hover:bg-blue-700 focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-blue-600 transition-colors cursor-pointer">
-                        <Rocket className="h-4 w-4"/>
-                        <span className="leading-none">Run (new to be integrate in the view page)</span>
-                    </button>
-                    <button
                         onClick={handleDataplayer}
                         aria-label={`View dataset for ${hit.title}`}
                         className="inline-flex items-center justify-center gap-1 rounded-md bg-blue-600 px-3 py-1.5 text-sm font-medium text-white shadow-sm hover:bg-blue-700 focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-blue-600 transition-colors cursor-pointer">
@@ -237,8 +215,8 @@ export const SearchResultItem = ({hit, isAiRanked = false}: SearchResultItemProp
                         <span className="leading-none">View</span>
                     </button>
                     <a href={hit._id} target="_blank" rel="noopener noreferrer"
-                       aria-label={`Redinect to the source of dataset ${hit.title}`}
-                       className="inline-flex items-center justify-center gap-1 rounded-md bg-green-600 px-3 py-1.5 text-sm font-medium text-white shadow-sm hover:bg-green-700 focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-green-600 transition-colors">
+                        aria-label={`Redinect to the source of dataset ${hit.title}`}
+                        className="inline-flex items-center justify-center gap-1 rounded-md bg-green-600 px-3 py-1.5 text-sm font-medium text-white shadow-sm hover:bg-green-700 focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-green-600 transition-colors">
                         <ExternalLinkIcon className="h-4 w-4"/>
                         <span className="leading-none">Source</span>
                     </a>
